requirements.txt
pillow==10.4.0
tkinterdnd2==0.4.3
ttkbootstrap==1.13.11
